refactor websockets package
