+++ internal/websockets/media.go
@@ -1,37 +1,8 @@
 package websockets
 
 import (
-	"github.com/slugger7/exorcist/internal/db/exorcist/public/model"
 	"github.com/slugger7/exorcist/internal/dto"
-	"github.com/slugger7/exorcist/internal/environment"
-	"github.com/slugger7/exorcist/internal/logger"
-	"github.com/slugger7/exorcist/internal/models"
 )
-
-type Websockets interface {
-	MediaOverviewUpdate(media dto.MediaOverviewDTO)
-	MediaUpdate(media dto.MediaDTO)
-	MediaDelete(media dto.MediaOverviewDTO)
-	MediaCreate(media dto.MediaOverviewDTO)
-	JobUpdate(job model.Job)
-}
-
-type websockets struct {
-	env    *environment.EnvironmentVariables
-	wss    models.WebSocketMap
-	logger logger.Logger
-}
-
-// JobUpdate implements Websockets.
-func (w *websockets) JobUpdate(job model.Job) {
-	w.logger.Debug("ws - updating job")
-
-	jobUpdate := dto.WSMessage[dto.JobDTO]{
-		Topic: dto.WSTopic_JobUpdate,
-		Data:  *(&dto.JobDTO{}).FromModel(job),
-	}
-	jobUpdate.SendToAll(w.wss)
-}
 
 // MediaCreate implements Websockets.
 func (w *websockets) MediaCreate(media dto.MediaOverviewDTO) {
@@ -77,18 +48,4 @@ func (w *websockets) MediaOverviewUpdate(media dto.MediaOverviewDTO) {
 		Data:  media,
 	}
 	mediaUpdate.SendToAll(w.wss)
-}
-
-var websocketsInterface *websockets
-
-func New(env *environment.EnvironmentVariables, wss models.WebSocketMap) Websockets {
-	if websocketsInterface == nil {
-		websocketsInterface = &websockets{
-			env:    env,
-			wss:    wss,
-			logger: logger.New(env),
-		}
-	}
-
-	return websocketsInterface
 }
